Description: Refactor resources and provider into a aiven module ready for import
 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	renamed:    kafka_topic_change.go -> aiven/kafka_topic_change.go
	renamed:    kafka_topic_create.go -> aiven/kafka_topic_create.go
	renamed:    provider.go -> aiven/provider.go
	renamed:    resource_connection_pool.go -> aiven/resource_connection_pool.go
	renamed:    resource_database.go -> aiven/resource_database.go
	renamed:    resource_kafka_acl.go -> aiven/resource_kafka_acl.go
	renamed:    resource_kafka_topic.go -> aiven/resource_kafka_topic.go
	renamed:    resource_project.go -> aiven/resource_project.go
	renamed:    resource_project_user.go -> aiven/resource_project_user.go
	renamed:    resource_project_vpc.go -> aiven/resource_project_vpc.go
	renamed:    resource_service.go -> aiven/resource_service.go
	renamed:    resource_service_integration.go -> aiven/resource_service_integration.go
	renamed:    resource_service_integration_endpoint.go -> aiven/resource_service_integration_endpoint.go
	renamed:    resource_service_user.go -> aiven/resource_service_user.go
	renamed:    resource_vpc_peering_connection.go -> aiven/resource_vpc_peering_connection.go
	renamed:    service_change.go -> aiven/service_change.go
	renamed:    templates/integration_endpoints_user_config_schema.json -> aiven/templates/integration_endpoints_user_config_schema.json
	renamed:    templates/integrations_user_config_schema.json -> aiven/templates/integrations_user_config_schema.json
	renamed:    templates/service_user_config_schema.json -> aiven/templates/service_user_config_schema.json
	renamed:    user_config.go -> aiven/user_config.go
	renamed:    version.go -> aiven/version.go
	modified:   go.mod
	modified:   go.sum
	modified:   main.go
