+++ aiven/resource_database.go
@@ -1,6 +1,6 @@
 // Copyright (c) 2017 jelmersnoeck
 // Copyright (c) 2018 Aiven, Helsinki, Finland. https://aiven.io/
-package main
+package aiven
 
 import (
 	"fmt"
