+++ aiven/templates/integration_endpoints_user_config_schema.json
