+++ aiven/templates/integrations_user_config_schema.json
