+++ aiven/templates/service_user_config_schema.json
